requirements.txt
matplotlib==3.3.4
jupyter_dash==0.4.0
dash_core_components==1.15.0
dash_table==4.11.2
mibian==0.1.3
Unipath==1.1
Django==3.1.7
numpy==1.20.1
dash==1.19.0
pysftp==0.2.9
dj_database_url==0.5.0
dash_html_components==1.1.2
plotly==4.4.1
yfinance==0.1.59
pandas==1.1.5
dash_bootstrap_components==0.12.0
django_model_utils==4.1.1
djangorestframework==3.12.4
python-decouple==3.4
django_plotly_dash
dpd_static_support
whitenoise
scipy
django_plotly_dash
xgboost==0.90